fix(): bind window as first param of getConfirmation
When `create***History` runs, `getUserConfirmation` defaults to the
`getConfirmation` function, which expects the global `window` object as
its first parameter.  However, it is later invoked without passing
`window`, probably because the signature for the user-defined flavor in
`CreateBrowserHistoryOptions` does not expect `window` to be passed.

The simple fix is to bind `window` to the default `getConfirmation`
function as its first parameter.