chore(ownership): move stencil router to community
this is the first commit necessary to move the router to the community
model. it changes the name of the package to `@stencil-community/router`
(from `@stencil/router`).

this commit does not update the demo package, as the router has not been
published to the (npm) package registry prior to this commit landing.
once this commit has landed, the demo package will be updated