chore(types): specify @types/node version
this commit add @types/node to the dependencies manifest for the router.
when trying to build this package without an earlier version of these
types, compilation fails. the compilation fails due to @types/node being
hoisted from the np package, and the types being hoisted are
incompatible with the version of typescript that is used with v1 of
stencil